+++ resources/js/modules/install/components/InstallingScreen.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import {t} from '@craftcms/ui';
-  import {computed, onMounted} from 'vue';
+  import {computed, nextTick, onMounted, useTemplateRef, watch} from 'vue';
   import {usePost} from '@/common/composables/useFetch';
   import {install as installAction} from '@actions/InstallController';
 
@@ -29,8 +29,16 @@
     error.value instanceof Error ? error.value.message : String(error.value)
   );
 
-  onMounted(async () => {
-    await install(props.data);
+  const heading = useTemplateRef<HTMLElement>('heading');
+
+  onMounted(() => {
+    install(props.data);
+  });
+  watch([isLoading, isSuccess, isError], async ([loading, success, error]) => {
+    if (loading || success || error) {
+      await nextTick();
+      heading.value?.focus();
+    }
   });
 </script>
 
@@ -38,14 +46,14 @@
   <craft-pane class="max-w-[80ch] mx-auto">
     <template v-if="isLoading">
       <div class="content">
-        <h2>{{ t('Installing Craft CMS…') }}</h2>
+        <h1 ref="heading" tabindex="-1">{{ t('Installing Craft CMS…') }}</h1>
         <craft-spinner></craft-spinner>
       </div>
     </template>
 
     <template v-else-if="isSuccess">
       <div class="content">
-        <h2>{{ t('Craft is installed! 🎉') }}</h2>
+        <h1 ref="heading" tabindex="-1">{{ t('Craft is installed! 🎉') }}</h1>
         <div class="flex justify-center items-center">
           <craft-icon
             name="circle-check"
@@ -56,9 +64,9 @@
       </div>
     </template>
 
-    <template v-if="isError">
+    <template v-else-if="isError">
       <div class="content">
-        <h2>{{ t('Install failed 😞') }}</h2>
+        <h1 ref="heading" tabindex="-1">{{ t('Install failed 😞') }}</h1>
 
         <div
           class="text-left border border-red-500 rounded p-4 text-red-800 bg-red-50 font-mono text-xs"